practice2/requirements.txt
fastapi
uvicorn
numpy
pillow
pytest
httpx
python-multipart
ffmpeg-python
streamlit
shutil